+++ inst/doc/dataset.R
@@ -4,14 +4,23 @@ knitr::opts_chunk$set(
   comment = "#>"
 )
 
-## ----setup--------------------------------------------------------------------
+## ----setup, message=FALSE-----------------------------------------------------
 library(dataset)
 
-## ---- include=FALSE-----------------------------------------------------------
+## ----data-frame-example-------------------------------------------------------
+foo <- data.frame( x = c(1,2), y = c(3,4))
+attr(foo, "Title") <- "My Foo Object"
+attributes(foo)
+
+## ----iris---------------------------------------------------------------------
+dataset_title(iris_dataset)
+publisher(iris_dataset)
+
+## ---- include=FALSE, message=FALSE--------------------------------------------
 library(here)
 library(knitr)
 
-## ---- message=FALSE-----------------------------------------------------------
+## ----eurostat-example, message=FALSE------------------------------------------
 rd_e_gerdtot <- eurostat::get_eurostat('rd_e_gerdtot')
 head(rd_e_gerdtot)
 
@@ -21,11 +30,11 @@ dimensions(rd_e_gerdtot) <- c("geo", "time", "sectperf")
 ## -----------------------------------------------------------------------------
 measures(rd_e_gerdtot) <- "value"
 
-## -----------------------------------------------------------------------------
+## ----attributes---------------------------------------------------------------
 attributes_measures(rd_e_gerdtot) <- "unit"
 datacite(rd_e_gerdtot)
 
-## ---- include=TRUE------------------------------------------------------------
+## ----datacite, include=TRUE---------------------------------------------------
 toc <- eurostat::get_eurostat_toc()
 rd_e_gerdtot_reference <- toc[which(toc$code == "rd_e_gerdtot"),]